address hang's comment
